requirements.txt
streamlit==1.29.0
reportlab==4.0.7
plotly==5.18.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
openai==1.6.1
Pillow==10.1.0
openpyxl==3.1.2